# backend/src/api-zod/generated/types/savedRoadmap.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SkillPath AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Roadmap } from './roadmap';

export interface SavedRoadmap {
  id: number;
  courseId: number;
  degreeName: string;
  createdAt: Date;
  roadmap: Roadmap;
}