# backend/src/api-zod/generated/types/university.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SkillPath AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface University {
  id: number;
  name: string;
  shortName: string;
  location: string;
  foundedYear: number;
  logoColor: string;
  ranking: number;
  /** @nullable */
  description?: string | null;
}