# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SkillPath AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stories:
    get:
      operationId: listSuccessStories
      tags: [stories]
      summary: List success stories
      parameters:
        - name: degreeType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of success stories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SuccessStory"
  /stories/{id}:
    get:
      operationId: getSuccessStory
      tags: [stories]
      summary: Get a success story by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success story details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessStory"
...
  /saved/courses:
    get:
      operationId: listSavedCourses
      tags: [saved]
      summary: List saved courses
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Saved courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedCourse"
    post:
      operationId: saveCourse
      tags: [saved]
      summary: Save a course
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveCourseInput"
      responses:
        "201":
          description: Course saved
  /saved/courses/{courseId}:
    delete:
      operationId: unsaveCourse
      tags: [saved]
      summary: Remove saved course
      security:
        - bearerAuth: []
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Course unsaved
  /saved/universities:
    get:
      operationId: listSavedUniversities
      tags: [saved]
      summary: List saved universities
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Saved universities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedUniversity"
    post:
      operationId: saveUniversity
      tags: [saved]
      summary: Save a university
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveUniversityInput"
      responses:
        "201":
          description: University saved
  /saved/universities/{universityId}:
    delete:
      operationId: unsaveUniversity
      tags: [saved]
      summary: Remove saved university
      security:
        - bearerAuth: []
      parameters:
        - name: universityId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: University unsaved
  /roadmaps:
    get:
      operationId: listRoadmaps
      tags: [roadmaps]
      summary: List user roadmaps
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Saved roadmaps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedRoadmap"
  /dashboard/recent-searches:
    get:
      operationId: listRecentSearches
      tags: [dashboard]
      summary: List recent searches
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Recent searches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecentSearch"
    post:
      operationId: recordSearch
      tags: [dashboard]
      summary: Record a search
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordSearchInput"
      responses:
        "201":
          description: Search recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecentSearch"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    University:
      type: object
      required: [id, name, shortName, location, foundedYear, logoColor, ranking]
      properties:
        id:
          type: integer
        name:
          type: string
        shortName:
          type: string
        location:
          type: string
        foundedYear:
          type: integer
        logoColor:
          type: string
        ranking:
          type: integer
        description:
          type: ["string", "null"]
    Course:
      type: object
      required: [id, universityId, universityName, degreeName, faculty, degreeType, durationYears, minimumZScore, stream]
      properties:
        id:
          type: integer
        universityId:
          type: integer
        universityName:
          type: string
        universityShortName:
          type: string
        degreeName:
          type: string
        faculty:
          type: string
        degreeType:
          type: string
        durationYears:
          type: integer
        minimumZScore:
          type: number
        stream:
          type: string
        matchScore:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        officialMinimumZScore:
          type: ["number", "null"]
        predictedMinimumZScore:
          type: ["number", "null"]
        officialAcademicYear:
          type: ["string", "null"]
        predictedAcademicYear:
          type: ["string", "null"]
        confidence:
          type: ["string", "null"]
          enum: [high, medium, low, null]
        eligibility:
          type: ["string", "null"]
          enum: [likely, borderline, reach, unlikely, null]
        dataSource:
          type: ["string", "null"]
        district:
          type: ["string", "null"]
...
    SuccessStory:
      type: object
      required: [id, name, degreeName, universityName, graduationYear, currentPosition, careerJourney]
      properties:
        id:
          type: integer
        name:
          type: string
        degreeName:
          type: string
        universityName:
          type: string
        graduationYear:
          type: integer
        currentPosition:
          type: string
        summary:
          type: ["string", "null"]
        careerJourney:
          type: array
          items:
            type: string
        avatarColor:
          type: ["string", "null"]
...
    Roadmap:
      type: object
      required: [courseId, degreeName, years, afterGraduation]
      properties:
        courseId:
          type: integer
        degreeName:
          type: string
        years:
          type: array
          items:
            $ref: "#/components/schemas/RoadmapYear"
        afterGraduation:
          type: array
          items:
            $ref: "#/components/schemas/CareerMilestone"
    RoadmapYear:
      type: object
      required: [year, milestones]
      properties:
        year:
          type: integer
        milestones:
          type: array
          items:
            type: string
    CareerMilestone:
      type: object
      required: [timeframe, role]
      properties:
        timeframe:
          type: string
        role:
          type: string
...
    SaveCourseInput:
      type: object
      required: [courseId]
      properties:
        courseId:
          type: integer
    SaveUniversityInput:
      type: object
      required: [universityId]
      properties:
        universityId:
          type: integer
    SavedCourse:
      allOf:
        - $ref: "#/components/schemas/Course"
        - type: object
          properties:
            savedAt:
              type: string
              format: date-time
    SavedUniversity:
      allOf:
        - $ref: "#/components/schemas/University"
        - type: object
          properties:
            savedAt:
              type: string
              format: date-time
    SavedRoadmap:
      type: object
      required: [id, courseId, degreeName, createdAt, roadmap]
      properties:
        id:
          type: integer
        courseId:
          type: integer
        degreeName:
          type: string
        createdAt:
          type: string
          format: date-time
        roadmap:
          $ref: "#/components/schemas/Roadmap"
    RecordSearchInput:
      type: object
      required: [query]
      properties:
        query:
          type: string
        filters:
          type: object
          additionalProperties: true
    RecentSearch:
      type: object
      required: [id, query, createdAt]
      properties:
        id:
          type: integer
        query:
          type: string
        filters:
          type: ["object", "null"]
          additionalProperties: true
        createdAt:
          type: string
          format: date-time